Prechod na jinou metodu
- zrychleni detekce
- dynamicka aktualizacepozadi
-> treba si vic hrat s paramtrama.

+++ get_background.m
@@ -12,7 +12,7 @@ for i=1:round(nframes/n):nframes
     R = R + double(read(Obj,i));
 end
     
-R = uint8(R/n);
+R = R/n;
 close(h)
 
 end
